requirements.txt
django
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
psycopg2
python-dotenv
